+++ NEWS.md
@@ -8,6 +8,7 @@
   * Probability distributions for PSA were renamed, see `?distributions`.
   * `state_cycle` renamed to `state_time`, `model_time` alias to `markov_cycle`.
   * The `inflow` argument for budget impact analysis is specified with `define_inflow()`.
+  * `prob_to_prob()`=>`rescale_prob()`.
 
 ## New features